move mock rule
